# client_app/openapi.yaml
{"openapi":"3.1.0","info":{"title":"Библиотека","version":"0.1.0"},"paths":{"/auth/jwt/login":{"post":{"tags":["auth"],"summary":"Auth:Jwt.Login","operationId":"auth_jwt_login_auth_jwt_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_auth_jwt_login_auth_jwt_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BearerResponse"},"example":{"access_token":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoiOTIyMWZmYzktNjQwZi00MzcyLTg2ZDMtY2U2NDJjYmE1NjAzIiwiYXVkIjoiZmFzdGFwaS11c2VyczphdXRoIiwiZXhwIjoxNTcxNTA0MTkzfQ.M10bjOe45I5Ncu_uXvOmVV8QxnL-nZfcH96U90JaocI","token_type":"bearer"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"LOGIN_BAD_CREDENTIALS":{"summary":"Bad credentials or the user is inactive.","value":{"detail":"LOGIN_BAD_CREDENTIALS"}},"LOGIN_USER_NOT_VERIFIED":{"summary":"The user is not verified.","value":{"detail":"LOGIN_USER_NOT_VERIFIED"}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/jwt/logout":{"post":{"tags":["auth"],"summary":"Auth:Jwt.Logout","operationId":"auth_jwt_logout_auth_jwt_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Missing token or inactive user."}},"security":[{"OAuth2PasswordBearer":[]}]}},"/auth/register":{"post":{"tags":["auth"],"summary":"Register:Register","operationId":"register_register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"REGISTER_USER_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"REGISTER_USER_ALREADY_EXISTS"}},"REGISTER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"REGISTER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"tags":["users"],"summary":"Users:Current User","operationId":"users_current_user_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."}},"security":[{"OAuth2PasswordBearer":[]}]},"patch":{"tags":["users"],"summary":"Users:Patch Current User","operationId":"users_patch_current_user_users_me_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"UPDATE_USER_EMAIL_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"UPDATE_USER_EMAIL_ALREADY_EXISTS"}},"UPDATE_USER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"UPDATE_USER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/{id}":{"get":{"tags":["users"],"summary":"Users:User","operationId":"users_user_users__id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["users"],"summary":"Users:Patch User","operationId":"users_patch_user_users__id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"400":{"content":{"application/json":{"examples":{"UPDATE_USER_EMAIL_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"UPDATE_USER_EMAIL_ALREADY_EXISTS"}},"UPDATE_USER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"UPDATE_USER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}},"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Bad Request"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["users"],"summary":"Users:Delete User","operationId":"users_delete_user_users__id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"204":{"description":"Successful Response"},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/take_book/{book_id}":{"post":{"summary":"Take Book","operationId":"take_book_take_book__book_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete_book/{book_id}":{"delete":{"summary":"Delete Book","operationId":"delete_book_delete_book__book_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/get_books_on_basket":{"get":{"summary":"Get Books On Basket","operationId":"get_books_on_basket_get_books_on_basket_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/books/":{"get":{"tags":["books"],"summary":"Get All Books","operationId":"get_all_books_books__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BookRead"},"type":"array","title":"Response Get All Books Books  Get"}}}}}},"post":{"tags":["books"],"summary":"Create New Book","operationId":"create_new_book_books__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookDB"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/{book_id}":{"get":{"tags":["books"],"summary":"Get All Books","operationId":"get_all_books_books__book_id__get","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["books"],"summary":"Delete Books","operationId":"delete_books_books__book_id__delete","parameters":[{"name":"book","in":"query","required":true,"schema":{"type":"integer","title":"Book"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["books"],"summary":"Update Book","operationId":"update_book_books__book_id__patch","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookDB"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/authors/":{"get":{"tags":["authors"],"summary":"Get All Author","operationId":"get_all_author_authors__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AuthorReadDB"},"type":"array","title":"Response Get All Author Authors  Get"}}}}}},"post":{"tags":["authors"],"summary":"Create New Author","operationId":"create_new_author_authors__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/authors/{author_id}":{"delete":{"tags":["authors"],"summary":"Delete Author","operationId":"delete_author_authors__author_id__delete","parameters":[{"name":"author_id","in":"path","required":true,"schema":{"type":"integer","title":"Author Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["authors"],"summary":"Update Author","operationId":"update_author_authors__author_id__patch","parameters":[{"name":"author_in","in":"query","required":true,"schema":{"type":"integer","title":"Author In"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorReadDB"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/authors/{book_id}/add_author/{author_id}":{"patch":{"tags":["authors"],"summary":"Update Book Author","description":"Добавление только одного автора в книгу","operationId":"update_book_author_authors__book_id__add_author__author_id__patch","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}},{"name":"author_id","in":"path","required":true,"schema":{"type":"integer","title":"Author Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/authors/{book_id}/delete_author/{author_id}":{"delete":{"tags":["authors"],"summary":"Update Book For Author","description":"Удаление автора из книги","operationId":"update_book_for_author_authors__book_id__delete_author__author_id__delete","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}},{"name":"author_id","in":"path","required":true,"schema":{"type":"integer","title":"Author Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AuthorBase":{"properties":{"name":{"type":"string","maxLength":100,"minLength":5,"title":"Name"},"biography":{"type":"string","maxLength":100,"minLength":3,"title":"Biography"},"birth_data":{"type":"string","format":"date","title":"Birth Data"}},"type":"object","required":["name","biography","birth_data"],"title":"AuthorBase"},"AuthorReadDB":{"properties":{"name":{"type":"string","maxLength":100,"minLength":5,"title":"Name"},"biography":{"type":"string","maxLength":100,"minLength":3,"title":"Biography"},"birth_data":{"type":"string","format":"date","title":"Birth Data"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","biography","birth_data","id"],"title":"AuthorReadDB"},"AuthorUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":100,"minLength":5},{"type":"null"}],"title":"Name"},"biography":{"anyOf":[{"type":"string","maxLength":100,"minLength":3},{"type":"null"}],"title":"Biography"},"birth_data":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Birth Data"}},"type":"object","title":"AuthorUpdate"},"BearerResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"BearerResponse"},"Body_auth_jwt_login_auth_jwt_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_auth_jwt_login_auth_jwt_login_post"},"BookDB":{"properties":{"name":{"type":"string","maxLength":100,"minLength":5,"title":"Name"},"description":{"type":"string","maxLength":100,"minLength":5,"title":"Description"},"public_data":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Public Data"},"book_genre":{"type":"integer","maximum":8.0,"minimum":1.0,"title":"Book Genre"},"book_copies":{"type":"integer","maximum":5.0,"minimum":0.0,"title":"Book Copies"},"book_data_ouput":{"type":"integer","maximum":60.0,"minimum":1.0,"title":"Book Data Ouput"},"authors":{"items":{"type":"integer"},"type":"array","title":"Authors"}},"type":"object","required":["name","description","book_genre","book_copies","authors"],"title":"BookDB"},"BookRead":{"properties":{"name":{"type":"string","maxLength":100,"minLength":5,"title":"Name"},"description":{"type":"string","maxLength":100,"minLength":5,"title":"Description"},"public_data":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Public Data"},"book_genre":{"type":"integer","maximum":8.0,"minimum":1.0,"title":"Book Genre"},"book_copies":{"type":"integer","maximum":5.0,"minimum":0.0,"title":"Book Copies"},"book_data_ouput":{"type":"integer","maximum":60.0,"minimum":1.0,"title":"Book Data Ouput"},"id":{"type":"integer","title":"Id"},"authors":{"items":{"$ref":"#/components/schemas/AuthorBase"},"type":"array","title":"Authors"}},"type":"object","required":["name","description","book_genre","book_copies","id","authors"],"title":"BookRead"},"ErrorModel":{"properties":{"detail":{"anyOf":[{"type":"string"},{"additionalProperties":{"type":"string"},"type":"object"}],"title":"Detail"}},"type":"object","required":["detail"],"title":"ErrorModel"},"GanreBook":{"type":"integer","enum":[1,2,3,4,5,6,7,8],"title":"GanreBook"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active","default":true},"is_superuser":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Superuser","default":false},"is_verified":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Verified","default":false},"first_name":{"type":"string","maxLength":100,"minLength":2,"title":"First Name"},"last_name":{"type":"string","maxLength":100,"minLength":2,"title":"Last Name"},"birth_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Birth Date"},"like_ganre":{"anyOf":[{"$ref":"#/components/schemas/GanreBook"},{"type":"null"}]},"telegram_id":{"type":"integer","title":"Telegram Id"}},"type":"object","required":["email","password","first_name","last_name","like_ganre","telegram_id"],"title":"UserCreate"},"UserRead":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","format":"email","title":"Email"},"is_active":{"type":"boolean","title":"Is Active","default":true},"is_superuser":{"type":"boolean","title":"Is Superuser","default":false},"is_verified":{"type":"boolean","title":"Is Verified","default":false},"first_name":{"type":"string","maxLength":100,"minLength":2,"title":"First Name"},"last_name":{"type":"string","maxLength":100,"minLength":2,"title":"Last Name"},"birth_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Birth Date"},"like_ganre":{"anyOf":[{"$ref":"#/components/schemas/GanreBook"},{"type":"null"}]},"telegram_id":{"type":"integer","title":"Telegram Id"}},"type":"object","required":["id","email","first_name","last_name","like_ganre","telegram_id"],"title":"UserRead"},"UserUpdate":{"properties":{"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"},"is_superuser":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Superuser"},"is_verified":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Verified"},"first_name":{"type":"string","maxLength":100,"minLength":2,"title":"First Name"},"last_name":{"type":"string","maxLength":100,"minLength":2,"title":"Last Name"},"birth_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Birth Date"}},"type":"object","required":["first_name","last_name"],"title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/jwt/login"}}}}}}